Added csString methods PadLeft(), PadRight(), PadCenter().  These pad the object to a specified length with a specified [or default space] char.
Added csString methods AsPadLeft(), AsPadRight(), AsPadCenter().  These return new strings formatted respectively with the aforementioned methods.

Added static cover methods PadLeft(), PadRight(), PadCenter() for all the common variants.

Intend to similarly implement Format().

Changed scfString and csString to take size_t instead of int for the reserved-length constructor.

Tested (though not yet exhaustively) on gnu/linux (gcc) and win32 (msvc).  Thanks to those who offered suggestions and feedback.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7940 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/scfstr.h
@@ -36,7 +36,7 @@ public:
   { SCF_CONSTRUCT_IBASE (NULL); }
 
   /// Create an scfString object and reserve space for iLength characters
-  scfString (int iLength) : s(iLength)
+  scfString (size_t iLength) : s(iLength)
   { SCF_CONSTRUCT_IBASE (NULL); }
 
   /// Copy constructor
